+++ batch.go
@@ -77,7 +77,7 @@ func (c *Client) CreateAndBuyBatchWithContext(ctx context.Context, in ...*Shipme
 
 // ListBatchesResult holds the results from the list insurances API.
 type ListBatchesResult struct {
-	Insurances []*Insurance `json:"batches,omitempty"`
+	Batch []*Batch `json:"batches,omitempty"`
 	// HasMore indicates if there are more responses to be fetched. If True,
 	// additional responses can be fetched by updating the ListInsurancesOptions
 	// parameter's AfterID field with the ID of the last item in this object's