+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 * Fix the bug for `AddShipmentsToBatch`, `AddShipmentsToBatchWithContext`, change the parameter from `string` to `interface{}` so user can pass in either Shipment IDs or Shipment structs
 * Fix the bug for `RemoveShipmentsFromBatch`, and `RemoveShipmentsFromBatchWithContext`, change the parameter from `*Shipment` to `interface{}` so user can pass in either Shipment IDs or Shipment structs
 * Add `CreateRefund`, `CreateRefundWithContext`, `ListRefunds`, `ListRefundsWithContext`, `GetRefund`, and `GetRefundWithContext` in the Refund file
+* Fix a typo in the `Batch` file, the type of `ListBatchesResult` should be `batch` instead of `insurance`.
+* Add missing `ID` attribute to `Brand`struct
 
 ## v2.0.1 (2022-02-10)